ConfirmDelete component

+++ src/Components/Pages/Admin/EventDelete/EventDelete.jsx
@@ -1,18 +1,19 @@
-import { useContext, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import "./EventDelete.css";
 import { FaTrashAlt } from "react-icons/fa";
 import CurrentEvent from "../../../../Context/CurrentEventContext";
 import { api } from "../../../../API/client";
 import ConfirmDelete from "./ConfirmDelete/ConfirmDelete";
+import ClearScreen from "../../../../Context/ClearScreen";
 
 const EventDelete = () => {
-  const [confirmDelete, setConfirmDelete] = useState(false);
   const { currentEvent, setCurrentEvent } = useContext(CurrentEvent);
+  const { clearScreen, setClearScreen } = useContext(ClearScreen);
 
   const handleDelete = () => {
     console.log(currentEvent);
-    setConfirmDelete(true);
     // deleteEvent();
+    setClearScreen({ ...clearScreen, btnEventAdd: true });
   };
 
   const deleteEvent = async () => {
@@ -21,9 +22,12 @@ const EventDelete = () => {
     console.log(resp);
   };
 
+  if (clearScreen.btnEventAdd) {
+    return <ConfirmDelete />;
+  }
+
   return (
     <div className="EventDelete" onClick={(e) => handleDelete(e)}>
-      {confirmDelete && <ConfirmDelete />}
       <FaTrashAlt className="trash" size={25} /> &nbsp; מחיקת אירוע
     </div>
   );
